Move `PartialVertex` to new `partial` module
