+++ crates/fj-kernel/src/lib.rs
@@ -91,5 +91,6 @@ pub mod algorithms;
 pub mod builder;
 pub mod iter;
 pub mod objects;
+pub mod partial;
 pub mod path;
 pub mod stores;
